skip if file output exists

+++ build/Publish.MacOS.cs
@@ -9,13 +9,13 @@ public partial class Build : FalloutBuild
 	{
         AbsolutePath virtualbackendBuildRoot = $"{GitRoot}/backend/virtualdisplay/.build";
 
-        if (File.Exists($"{virtualbackendBuildRoot}/out/Products/Release/libvirtualdisplay.dylib"))
+        if (!File.Exists((AbsolutePath)outputDirectory / "libvirtualdisplay.dylib")
+            && File.Exists($"{virtualbackendBuildRoot}/out/Products/Release/libvirtualdisplay.dylib"))
         {
             AbsolutePath source = $"{virtualbackendBuildRoot}/out/Products/Release/libvirtualdisplay.dylib";
             Log.Information("Copying over {Source} into {OutDir}", source, OutputRoot);
             File.Copy(source, (AbsolutePath)outputDirectory / "libvirtualdisplay.dylib");
-        }
-        if (File.Exists($"{virtualbackendBuildRoot}/apple/Products/Release/libvirtualdisplay.dylib"))
+        } else if (File.Exists($"{virtualbackendBuildRoot}/apple/Products/Release/libvirtualdisplay.dylib"))
         {
             AbsolutePath source = $"{virtualbackendBuildRoot}/apple/Products/Release/libvirtualdisplay.dylib";
             Log.Information("Copying over {Source} into {OutDir}", source, OutputRoot);
